Server/bot/requirements.txt
sqlalchemy
psycopg2-binary
python-telegram-bot